+++ src/jsdoc-globals.js
@@ -263,6 +263,9 @@ var FancytreeOptions = {};
  * @property {function} createNode Allow tweaking and binding, after node was created for the first time (NOTE: this event is only available as callback, but not for bind())
  * @property {function} dblclick `data.node` was double-clicked. `data.targetType` contains the region ("checkbox", "expander", "icon", "prefix", "title"). Return `false` to prevent default processing, i.e. expanding, etc.
  * @property {function} deactivate `data.node` was deactivated
+ * @property {function} defaultGridAction (used by ext-aria) The user hit enter on the active row or cell.<br>
+ *     `data.activeTd` contains the currently active &lt;td> element or null<br>
+ *     `data.colIdx` contains the 0-based column index or -1
  * @property {function} enhanceTitle Allow extending the `&lt;span class='fancytree-title'>` markup, for example by adding badges, ... (NOTE: this event is only available as callback, but not for bind())
  * @property {function} expand `data.node` was expanded
  * @property {function} focus `data.node` received keyboard focus