+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 2.29.1-0 / Unreleased
-  *
+  * Fix ES6 import dependency on jquery for jquery.fancytree.ui-deps.js
 
 # 2.29.0 / 2018-06-16
   * [Changed]