+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * [Fixed] ES6 import dependency on jquery for jquery.fancytree.ui-deps.js
   * [Fixed] #848 Drag End Error with dnd5 extension (again):
     fancytree-drag-remove class not removed on drop/dragend
+  * [Fixed] ext-ariagrid default actions
 
 # 2.29.0 / 2018-06-16
   * [Changed]